refactor(router): separate client responsibilities
Split Link, navigation, and render-time error handling into focused client modules.

Keep the public API and behavior unchanged.
